+++ NEWS.md
@@ -1,3 +1,19 @@
+DatabaseConnector 6.3.0
+=======================
+
+Changes:
+
+1. On Snowflake always using `QUOTED_IDENTIFIERS_IGNORE_CASE=TRUE` to avoid name mismatches when using quotes.
+
+2. Updated Redshift drivers.
+
+3. Added unit tests for all supported platforms.
+
+Bugfixes:
+
+1. Fix bug on BigQuery where wait time was too short to avoid rate limit error.
+
+
 DatabaseConnector 6.2.4
 =======================